Flask_application/app/requirements.txt
Flask==1.0.2
numpy==1.15.4
requests==2.20.1
ujson==1.35
tqdm==4.41.1
tensorflow==2.4.1
tensorboard==2.4.1
pytesseract==0.3.7
Pillow==7.0.0
msgpack==1.0.2

